% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_subcategories}
\alias{clean_subcategories}
\title{Cleaning subcategorical values}
\usage{
clean_subcategories(values)
}
\arguments{
\item{values}{values from get_val_data function}
}
\value{
something cool
}
\description{
Cleaning subcategorical values
}
