% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_query}
\alias{create_query}
\title{Create query for queryList}
\usage{
create_query(vars, vals)
}
\arguments{
\item{vars}{variables}

\item{vals}{values}
}
\value{
the "queryList" query
}
\description{
Create query for queryList
}
