% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_px}
\alias{get_px}
\title{Get px from hagstofan or other px sources}
\usage{
get_px(url, queryList)
}
\arguments{
\item{url}{the url to a json px source}

\item{queryList}{a list of data from px source}
}
\value{
something neat
}
\description{
Get px from hagstofan or other px sources
}
