% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_val_data}
\alias{get_val_data}
\title{Get factor values from variable}
\usage{
get_val_data(url, variable)
}
\arguments{
\item{url}{the url to a json px source}

\item{variable}{the name of a variable i.e. a column variable}
}
\value{
example data
}
\description{
Get factor values from variable
}
