% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_var_data}
\alias{get_var_data}
\title{Clean the json from Hagstofan and extract either example values or example variables}
\usage{
get_var_data(url, identity)
}
\arguments{
\item{url}{the url to a json px source}

\item{identity}{identity of data needed to extract ('value' or 'code')}
}
\value{
example data
}
\description{
Clean the json from Hagstofan and extract either example values or example variables
}
