% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_func.R
\name{hackstofan}
\alias{hackstofan}
\title{Get data from Hagstofan json px based on one known variable}
\usage{
hackstofan(url, variable)
}
\arguments{
\item{url}{the url to a json px source}

\item{variable}{known variable from the data}
}
\value{
something cool
}
\description{
Get data from Hagstofan json px based on one known variable
}
\examples{
\dontrun{
# url <- some http url from px.hagstofa.is or other px json website
# variable <- name of one variable in the px json
data <- hackstofan(url, variable)
}
}
